+++ System.Net.Mqtt/Packets/V3/UnsubscribePacket.cs
@@ -85,7 +85,7 @@ public sealed class UnsubscribePacket : MqttPacketWithId
             remainingLength += filters[i].Length + 2;
         }
 
-        var size = 1 + MqttExtensions.GetVarBytesCount(remainingLength) + remainingLength;
+        var size = 1 + MqttExtensions.GetVarBytesCount((uint)remainingLength) + remainingLength;
         var span = buffer = writer.GetSpan(size);
 
         span[0] = UnsubscribeMask;
